move things around to read easier
lets, vars, extensions, classes - alphabetized
