+++ Nudge/Utilities/Logger.swift
@@ -8,7 +8,6 @@
 import Foundation
 import os
 
-let bundleID = bundle.bundleIdentifier ?? "com.github.macadmins.Nudge"
 let utilsLog = Logger(subsystem: bundleID, category: "utilities")
 let osLog = Logger(subsystem: bundleID, category: "operating-system")
 let loggingLog = Logger(subsystem: bundleID, category: "logging")
